api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
